Move callRPCMethod method to helpers file
